renamed the package fro MyPackage to BicoccaCoursePython2024
