+++ Esercizi/BicoccaCoursePython2024/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = MyPackage
-version = 0.1.0
+name = BicoccaCoursePython2024
+version = 0.0.1
 description = Libreria per il corso di scientific python di Gerosa
 author = Francesco Turini
 author_email = fturini.turini7@gmail.com
@@ -18,6 +18,8 @@ classifiers =
     Intended Audience :: Science/Research
     Topic :: Scientific/Engineering
 
+url = https://github.com/fturini98/scientificcomputing_bicocca_2024.git@Esercizi-FrancescoTurini
+
 [options]
 package_dir =
     =src
@@ -27,6 +29,7 @@ install_requires =
     numpy
     scipy
     matplotlib
+    numba
 python_requires = >=3.7
 
 [options.packages.find]
