+++ Esercizi/BicoccaCoursePython2024/src/BicoccaCoursePython2024/__init__.py
@@ -1,0 +1,1 @@
+print('Paccetto BicoccaCoursePython2024 importato correttamente')
