+++ Esercizi/BicoccaCoursePython2024/src/BicoccaCoursePython2024/prima_lezione.py
