+++ Esercizi/BicoccaCoursePython2024/src/BicoccaCoursePython2024/quarta_lezione.py
