+++ Esercizi/BicoccaCoursePython2024/src/BicoccaCoursePython2024/seconda_lezione.py
