+++ Esercizi/BicoccaCoursePython2024/test/PoP10_GameOfLife.npy
