--- Esercizi/MyPackage/src/MyPackage/__init__.py
@@ -1,1 +1,0 @@
-print('Paccetto MyPackage importato correttamente')
